Fixed MSVC warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3552 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/AuxLib.cpp
@@ -434,7 +434,7 @@ void CapFPS() {
 
 	// Cap the FPS
 	if(currentTime - tLX->currentTime < fMaxFrameTime)
-		SDL_Delay((fMaxFrameTime - (currentTime - tLX->currentTime)).milliseconds());
+		SDL_Delay((Uint32)(fMaxFrameTime - (currentTime - tLX->currentTime)).milliseconds());
 	else
 		// do at least one small break, else it's possible that we never receive signals from our OS
 		SDL_Delay(1);
